pyproject.toml
[project]
name = "circt-adapter"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.12, <=3.12.8"
dependencies = [
  "ai-edge-model-explorer",

  # uv add -r circt/llvm/mlir/python/requirements.txt
  "ml-dtypes>=0.1.0,<=0.6.0 ; python_full_version < '3.13'",
  "ml-dtypes>=0.5.0,<=0.6.0 ; python_full_version >= '3.13'",
  "nanobind>=2.4,<3.0",
  "numpy>=1.19.5,<=2.1.2",
  "pybind11>=2.10.0,<=2.13.6",
  "pyyaml>=5.4.0,<=6.0.1",
]
